refactor: [code-review] Split `job_run_store/mod.rs` and move its inlin… [ORB-11653]
[code-review] Split `job_run_store/mod.rs` and move its inline tests to the sibling layout

Planned-By: grok
Orchestrated-By: astra
